requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
openai==0.26.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==1.7.1
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.14
yarl==1.8.2
youtube-search-python==1.6.6
gunicorn==20.1.0